Move Gmail CSS class definitions to a new module
Allow them to be used for a new Sidebar module that will enable sidebar
toggling.

Had to switch the compiled module system to `amd` because of this error
message:

    tsconfig.json:3:3 - error TS6082: Only 'amd' and 'system' modules
    are supported alongside --outFile.

Thinking I'm going to switch to Typescript namespaces to save myself the
trouble of using a 'require' library for the AMD modules.