Try to add 'j/k' shortcuts in the pager view
Very messy at the moment. Doesn't currently work because it needs to be
executed when the pager view becomes visible. Pager divs are created and
removed each time a message is opened.

Going to put this aside for now and try to focus on the index view. From
there, I should be able to figure out how to load in pager commands when
a message is selected.